cloud/python/src/face_descriptor/requirements.txt
click==8.1.7
dlib==19.24.2
dnspython==2.4.2
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.24.4
opencv-python==4.8.1.78
Pillow==10.0.1
pymongo==4.5.0
